requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
bcrypt==5.0.0
bitarray==3.7.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
ckzg==2.1.5
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
cytoolz==1.1.0
ecdsa==0.19.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.1
eth_abi==5.2.0
fastapi==0.120.2
frozenlist==1.8.0
h11==0.16.0
hexbytes==1.3.1
idna==3.11
multidict==6.7.0
parsimonious==0.10.0
propcache==0.4.1
pyasn1==0.6.1
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pyunormalize==17.0.0
pywin32==311
regex==2025.10.23
requests==2.32.5
rlp==4.1.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.49.1
toolz==1.1.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
web3==7.14.0
websockets==15.0.1
yarl==1.22.0
